Splits app.py into modularized files
